feat: add draggable attribute
