+++ docs/en-US/component/input-tag.md
@@ -57,6 +57,16 @@ input-tag/clearable
 
 :::
 
+## Draggable
+
+You can set whether tags can be dragged.
+
+:::demo
+
+input-tag/draggable
+
+:::
+
 ## Sizes
 
 Add `size` attribute to change the size of InputTag. In addition to the default size, there are two other options: `large`, `small`.
@@ -91,25 +101,26 @@ input-tag/prefix-suffix
 
 ### Attributes
 
-| Name                  | Description                                                            | Type                                                        | Default |
-| --------------------- | ---------------------------------------------------------------------- | ----------------------------------------------------------- | ------- |
-| model-value / v-model | binding value                                                          | ^[array]`string[]`                                          | —       |
-| max                   | max number tags that can be enter                                      | ^[number]                                                   | —       |
-| trigger               | the key to trigger input tag                                           | ^[enum]`'Enter' \| 'Space'`                                 | Enter   |
-| tag-type              | tag type                                                               | ^[enum]`'' \| 'success' \| 'info' \| 'warning' \| 'danger'` | info    |
-| tag-effect            | tag effect                                                             | ^[enum]`'' \| 'light' \| 'dark' \| 'plain'`                 | light   |
-| size                  | input box size                                                         | ^[enum]`'large' \| 'default' \| 'small'`                    | —       |
-| clearable             | whether to show clear button, only works when `type` is not 'textarea' | ^[boolean]                                                  | false   |
-| disabled              | whether Input is disabled                                              | ^[boolean]                                                  | false   |
-| readonly              | same as `readonly` in native input                                     | ^[boolean]                                                  | false   |
-| autofocus             | same as `autofocus` in native input                                    | ^[boolean]                                                  | false   |
-| validate-event        | whether to trigger form validation                                     | ^[boolean]                                                  | true    |
-| id                    | native input id                                                        | ^[string]                                                   | —       |
-| tabindex              | input tabindex                                                         | ^[string] / ^[number]                                       | —       |
-| maxlength             | same as `maxlength` in native input                                    | ^[string] / ^[number]                                       | —       |
-| minlength             | same as `minlength` in native input                                    | ^[string] / ^[number]                                       | —       |
-| placeholder           | placeholder of Input                                                   | ^[string]                                                   | —       |
-| autocomplete          | same as `autocomplete` in native input                                 | ^[string]                                                   | off     |
+| Name                  | Description                            | Type                                                        | Default |
+| --------------------- | -------------------------------------- | ----------------------------------------------------------- | ------- |
+| model-value / v-model | binding value                          | ^[array]`string[]`                                          | —       |
+| max                   | max number tags that can be enter      | ^[number]                                                   | —       |
+| tag-type              | tag type                               | ^[enum]`'' \| 'success' \| 'info' \| 'warning' \| 'danger'` | info    |
+| tag-effect            | tag effect                             | ^[enum]`'' \| 'light' \| 'dark' \| 'plain'`                 | light   |
+| trigger               | the key to trigger input tag           | ^[enum]`'Enter' \| 'Space'`                                 | Enter   |
+| draggable             | whether tags can be dragged            | ^[boolean]                                                  | false   |
+| size                  | input box size                         | ^[enum]`'large' \| 'default' \| 'small'`                    | —       |
+| clearable             | whether to show clear button           | ^[boolean]                                                  | false   |
+| disabled              | whether to disable input-tag           | ^[boolean]                                                  | false   |
+| validate-event        | whether to trigger form validation     | ^[boolean]                                                  | true    |
+| readonly              | same as `readonly` in native input     | ^[boolean]                                                  | false   |
+| autofocus             | same as `autofocus` in native input    | ^[boolean]                                                  | false   |
+| id                    | same as `id` in native input           | ^[string]                                                   | —       |
+| tabindex              | same as `tabindex` in native input     | ^[string] / ^[number]                                       | —       |
+| maxlength             | same as `maxlength` in native input    | ^[string] / ^[number]                                       | —       |
+| minlength             | same as `minlength` in native input    | ^[string] / ^[number]                                       | —       |
+| placeholder           | placeholder of input                   | ^[string]                                                   | —       |
+| autocomplete          | same as `autocomplete` in native input | ^[string]                                                   | off     |
 
 ### Events
